+++ src/api/requests/libraryRequests.ts
@@ -1,4 +1,5 @@
-import { ApiResponseWithMeta, SearchLibraryReponse, SearchResponseMeta } from "../../Types"
+import { SearchLibraryReponse } from "../../types/LibraryTypes"
+import { ApiResponseWithMeta, SearchResponseMeta } from "../../types/UtilityTypes"
 import api from "../api"
 
 export const fetchLibraryAsync = async () => {
