+++ src/types/LibraryTypes.ts
@@ -1,0 +1,22 @@
+export interface Game {
+    id: string
+    title: string
+    status: string
+    console: string
+    format: string
+    genre: string
+    length: number
+    createdAt: Date
+    updatedAt: Date
+    dateCompleted: Date
+}
+
+export interface SearchLibraryReponse {
+    games: Game[]
+}
+
+export enum PanelStatus {
+    Viewing,
+    Adding,
+    Editing
+}
