+++ src/types/UtilityTypes.ts
@@ -7,23 +7,6 @@ export interface SerializeableAxiosResponse {
     statusCode: number
 }
 
-export interface Game {
-    id: string
-    title: string
-    status: string
-    console: string
-    format: string
-    genre: string
-    length: number
-    createdAt: Date
-    updatedAt: Date
-    dateCompleted: Date
-}
-
-export interface SearchLibraryReponse {
-    games: Game[]
-}
-
 export interface ApiResponse<TData> {
     data: TData
 }
@@ -35,10 +18,4 @@ export interface ApiResponseWithMeta<TMeta, TData> {
 
 export interface SearchResponseMeta {
     count: number
-}
-
-export enum PanelStatus {
-    Viewing,
-    Adding,
-    Editing
 }
